src/Notification/requirements.txt
wfastcgi==3.0.0
Django==5.0.12
djangorestframework==3.15.2
jalali_core==1.0.0
jdatetime==5.2.0
mssql-django==1.5
PyJWT==2.10.1
pyodbc==5.2.0
pytz==2025.1
tzdata==2025.1
django-mis-auth2==0.3.1
setuptools==75.8.0
python-dotenv==1.1.0
